pyproject.toml
[tool.poetry]
name = "nadoo-sandbox"
version = "0.1.0"
description = "Secure code execution service for Nadoo AI Platform"
authors = ["Nadoo Team <dev@nadoo.ai>"]
license = "MIT"
readme = "README.md"
homepage = "https://nadoo.ai"
repository = "https://github.com/nadoo-ai/sandbox"
documentation = "https://docs.nadoo.ai"
keywords = ["nadoo", "sandbox", "code-execution", "docker", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
docker = "^7.0.0"
aiofiles = "^23.2.1"
asyncio = "^3.4.3"
httpx = "^0.25.2"
redis = "^5.0.1"
celery = "^5.3.4"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.21.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
posthog = "^3.7.4"

# Cloud providers (optional)
boto3 = {version = "^1.34.0", optional = true}
google-cloud-run = {version = "^0.10.0", optional = true}
azure-mgmt-appcontainers = {version = "^3.0.0", optional = true}
azure-identity = {version = "^1.15.0", optional = true}

[tool.poetry.extras]
aws = ["boto3"]
gcp = ["google-cloud-run"]
azure = ["azure-mgmt-appcontainers", "azure-identity"]
cloud = ["boto3", "google-cloud-run", "azure-mgmt-appcontainers", "azure-identity"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
pythonpath = ["src"]
